# swagger.yaml
openapi: 3.0.3
info:
  title: Job-API
  contact: {}
  version: "1.0"
servers:
  - url: https://jobs-apis.onrender.com/api/v1
    variables: {}
paths:
...
  /jobs:
    get:
      tags:
        - Jobs
      summary: Get All Jobs
      operationId: GetAllJobs
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
    post:
      tags:
        - Jobs
      summary: Create Job
      operationId: CreateJob
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CreateJobRequest"
            example:
              company: Swagger Job
              position: Swagger Position
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
    parameters: []
...
components:
  schemas:
...
    CreateJobRequest:
      title: CreateJobRequest
      required:
        - company
        - position
      type: object
      properties:
        company:
          type: string
        position:
          type: string
      example:
        - company: New Job
          position: New Position
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...